+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-1. [#183](https://github.com/InfluxCommunity/influxdb3-js/pull/183): Add default tags
+1. [#183](https://github.com/InfluxCommunity/influxdb3-js/pull/183): Default Tags for Writes
 
 ## 0.4.1 [2023-11-16]